verify_receipt could never be called either — issue the key
`verify_receipt` does an exact `find()` on a full entry digest, and NO
surface in this system emitted one. `read_ledger` abbreviated through
`short()`; `current_phase` shortened the tip; `commit_to_round` sliced
its entry hash to twelve characters; the other write tools replied with
sequence numbers; every UI render abbreviated too. The tool worked
perfectly and its key was never issued.

Found by an independent audit pass asked to hunt one specific defect
class, and it is the fourth instance of that class in a day.

WHY THIS ONE MATTERED MORE THAN get_artefact. `verify_receipt` is the
ONLY tool-side surface that renders the confirmation fingerprint and the
carried partner claim. So both repairs that landed earlier today —
3bbf6e1 putting the fingerprint on the entry, c6d4efa recording the
claim it had promised to record verbatim — were invisible to any agent.
We fixed the recording twice and left the reading shut. A defect that
silently negates two prior fixes is worth more than its diff.

`read_ledger` now emits the full digest. `prev` stays abbreviated on
purpose: it is a back-reference, the preceding row carries that same
digest in full, nothing takes it as input, and repeating it would double
the cost against the 1500-char output budget. That budget still applies
— a 20-entry window will truncate — but truncation announces itself and
tells the caller to narrow the window, which is honest and, here,
actually actionable.

TWO ROUND-TRIP TESTS, and the second is the one that matters: it reads
the ledger as an agent must, takes a digest out of what it was handed,
verifies with it, and asserts the PARTNER'S CLAIM TEXT comes back. That
asserts the whole path rather than the storage, which is where the last
three defects hid.

NEGATIVE CONTROL, and a note on how nearly it went wrong. The first
attempt used `sed` and silently failed to match, so both tests "passed"
against unmodified code and would have been reported as controlled. The
grep that was supposed to confirm the revert returned 0 and I read it.
Re-run through a marker-asserting script: with the full hash reverted to
`short()`, both new tests FAIL; restored, all 26 pass. A control that
cannot be confirmed to have applied is not a control.

README's test count was stale at 23 against a suite of 24 — a checkable
claim, wrong, in the file GitHub renders first, two days before this
repo goes public. Now 26.

+++ src/tools.js
@@ -68,8 +68,18 @@ export function buildTools(room, { onCall } = {}) {
         const all = room.ledger.entries;
         const slice = all.slice(from - 1, from - 1 + n);
         if (!slice.length) return text(`No entries at cursor ${from}. Ledger holds ${all.length}.`);
+        // The entry hash in FULL. This listing is the only place an agent can
+        // ever learn one, and `verify_receipt` does an exact `find()` on a full
+        // digest — so abbreviating here made that tool permanently uncallable,
+        // and with it the ONLY tool-side view of the confirmation fingerprint
+        // and the carried partner claim. Two fixes that landed earlier today
+        // were invisible to agents because of this one call to `short()`.
+        //
+        // `prev` stays abbreviated on purpose: it is a back-reference, and the
+        // preceding row carries that same digest in full. Nothing needs it as
+        // an input, and repeating it would double the cost against the budget.
         const lines = slice.map(
-          (e) => `#${e.seq} ${e.kind} — ${e.actor.seat} via ${e.actor.ingress} [${e.actor.grade}]\n   ${short(e.hash)} prev ${short(e.prev)}`,
+          (e) => `#${e.seq} ${e.kind} — ${e.actor.seat} via ${e.actor.ingress} [${e.actor.grade}]\n   ${e.hash}  prev ${short(e.prev)}`,
         );
         const next = from + slice.length;
         return text(
